Fix requestLog (CU_REQUEST_LOG)
Note that this is currently disabled in production since it broke
when we moved to express 4, since it removes app.routes.

`req.route` is populated after next() is called, so we can use it here.

I'd like to add that it is slightly nasty that this is a thing which
isn't exercised during testing (because it's only used when
CU_REQUEST_LOG is set), and therefore meant we didn't catch this before
we tried to go live. Blegh.