Further cleaning to mocha wrapper
Mocha gained a `--recursive` parameter, so we don't need to do a find
anymore. Now the only thing left to figure out is how to run the
cleaner, then we can eliminate the wrapper entirely. It's ugly.

I'm thinking maybe we could add a --require to mocha.opts to do it, but
I'm not sure how to deal with the async.

Care to weigh in @drj11?