Update wd40 to 0.1.0
Previously the version specification was "*", which probably makes
sense in that we control wd40, so if we're updating it, it's probably
because we want to use it. OTOH, I want to actually document in the
custard repository when wd40 is changed in an important way.

I've published it to NPM.